Split tenosr.go
